Add the Parser code and change licence to GNU/GPL3
parser.js provide several functions to parse the language. You can test
if a text is valid to the language syntax and parse it to return a
formatted array. All the functions are unit tested. Run the test via
`npm test` in the project root directory.

Change licence from MIT to GNU/GPL v3.0